requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
environs==11.0.0
marshmallow==3.21.2
mysqlclient==2.2.4
packaging==24.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
